requirements.txt
yt-dlp
youtube-transcript-api
streamlit
python-dotenv
openai
openai-whisper
google-generativeai
